Implementando notificações de erros

+++ src/DevIO.Business/Services/SupplierService.cs
@@ -9,6 +9,12 @@ public class SupplierService(ISupplierRepository supplierRepository) : BaseServi
         if (!ExecuteValidation(new SupplierValidation(), supplier)
             || !ExecuteValidation(new AddressValidation(), supplier.Address)) return;
 
+        if (_supplierRepository.Search(s => s.Document == supplier.Document).Result.Any())
+        {
+            Notify("Already exists a supplier with this document.");
+            return;
+        }
+
         await _supplierRepository.AddAsync(supplier);
     }
 
@@ -16,11 +22,36 @@ public class SupplierService(ISupplierRepository supplierRepository) : BaseServi
     {
         if (!ExecuteValidation(new SupplierValidation(), supplier)) return;
 
+        if (_supplierRepository.Search(s => s.Document == supplier.Document && s.Id == supplier.Id).Result.Any())
+        {
+            Notify("Already exists a supplier with this document.");
+            return;
+        }
+
         await _supplierRepository.UpdateAsync(supplier);
     }
 
     public async Task RemoveAsync(Guid id)
     {
+        var supplier = await _supplierRepository.GetSupplerWithProductsAndAddressAsync(id);
+        if (supplier is null)
+        {
+            Notify("Supplier does not exist!");
+            return;
+        }
+
+        if (supplier.Products.Any())
+        {
+            Notify("The supplier has assciated products!");
+            return;
+        }
+
+        var address = await _supplierRepository.GetAddresBySupplierAsync(id);
+        if (address is not null)
+        {
+            await _supplierRepository.RemoveSupplierAddressAsync(address);
+        }
+
         await _supplierRepository.RemoveAsync(id);
     }
